backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.1
attrs==23.2.0
bcrypt==4.2.0
cachetools==5.3.3
certifi==2023.5.7
cffi==1.17.0
charset-normalizer==3.1.0
contourpy==1.1.1
cryptography==43.0.0
cycler==0.12.1
Django==4.2.1
django-cors-headers==4.4.0
django-debug-toolbar==4.1.0
djangorestframework==3.15.2
dnspython==2.0.0
fonttools==4.43.1
frozenlist==1.4.1
future==0.18.3
gunicorn==23.0.0
geoip2==4.8.0
idna==3.4
ipinfo==5.0.1
ipwhois==1.2.0
kiwisolver==1.4.5
Markdown==3.6
matplotlib==3.8.1
maxminddb==2.6.0
memory-profiler==0.61.0
multidict==6.0.5
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
psutil==5.9.6
pycountry==23.12.11
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.9.0
pymongo==4.4.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-whois==0.8.0
python3-nmap==1.6.0
pytz==2023.3
requests==2.31.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tldextract==5.1.3
urllib3==2.0.2
yarl==1.9.4
